feature/module Add ES module output (#219)
* Add ES module output

* .T should return copy

* Fix inverted contiguous bool

* Add exports field

* Update readme

* Add increments and checks for blas calls

* Generate dist

* Fix tolerance issue in rank function

* Contiguous flag causes unexpected issues, remove for now

* Update ci, add release workflow

* Fix bug in fill method

* LAPACK acceleration in gauss() not working, remove for now

* Fix bugs in test suite

* Remove dist before build

* Generate dist

* Automatically tag releases

+++ src/core/range.ts
@@ -21,7 +21,7 @@ import { NDArray } from './';
  * range(1, 2, 9); // => array([1, 3, 5, 7])
  */
 export const range = (...args: number[]): NDArray => {
-  const type: TypedArrayConstructor = Float32Array;
+  const type: TypedArrayConstructor = Float64Array;
   let backwards: boolean = false;
   let start: number;
   let step: number;
